+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 - Fixes
-  - Grid view takes int account header sizes when sizing columns
-  - Fixed grid view copy to clipboard actions
+  - Grid view takes into account header sizes when sizing columns
+  - Fixed grid view copy to clipboard context menu actions
 - General:
   - Cobol view subheaders context menus and size tooltip
   - Text view subheaders shows file name in context menu