+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Fixes
   - Grid view takes into account header sizes when sizing columns
   - Fixed grid view copy to clipboard context menu actions
+  - Fixed value not being updated in some case in MRU text combos
 - General:
   - Cobol open file dialog supports dropping files
   - Cobol view subheaders context menus and size tooltip